pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "pillow>=11.3.0",
    "requests>=2.32.4",
    "schedule>=1.2.2",
    "tweepy>=4.16.0",
]
